+++ core/src/test/java/nl/nn/adapterframework/management/bus/TestSecurityItems.java
@@ -7,9 +7,9 @@ import org.springframework.messaging.Message;
 
 import nl.nn.adapterframework.jms.JmsRealm;
 import nl.nn.adapterframework.jms.JmsRealmFactory;
-import nl.nn.adapterframework.testutil.FixedQuerySenderMock.ResultSetBuilder;
 import nl.nn.adapterframework.testutil.MatchUtils;
 import nl.nn.adapterframework.testutil.TestFileUtils;
+import nl.nn.adapterframework.testutil.mock.FixedQuerySenderMock.ResultSetBuilder;
 
 public class TestSecurityItems extends BusTestBase {
