+++ core/src/test/java/nl/nn/adapterframework/testutil/TestConfiguration.java
@@ -74,7 +74,7 @@ public class TestConfiguration extends Configuration {
 
 	/**
 	 * Add the ability to mock FixedQuerySender ResultSets. Enter the initial query and a mocked 
-	 * ResultSet using a {@link FixedQuerySenderMock.ResultSetBuilder ResultSetBuilder}.
+	 * ResultSet using a {@link nl.nn.adapterframework.testutil.mock.FixedQuerySenderMock.ResultSetBuilder ResultSetBuilder}.
 	 */
 	public void mockQuery(String query, ResultSet resultSet) {
 		qsPostProcessor.addMock(query, resultSet);
